Moved EducationGate to be shared by multiple forms
