+++ src/applications/edu-benefits/0994/Form0994App.jsx
@@ -2,7 +2,7 @@ import React from 'react';
 
 import RoutedSavableApp from 'platform/forms/save-in-progress/RoutedSavableApp';
 import formConfig from './config/form';
-import EducationGate from './containers/EducationGate';
+import EducationGate from '../containers/EducationGate';
 
 export default function Form0994Entry({ location, children }) {
   return (
