+++ src/applications/edu-benefits/0994/tests/containers/EducationGate.unit.spec.jsx
@@ -2,7 +2,7 @@ import React from 'react';
 import { expect } from 'chai';
 import { shallow } from 'enzyme';
 
-import { EducationGate } from '../../../0994/containers/EducationGate';
+import { EducationGate } from '../../../containers/EducationGate';
 import backendServices from 'platform/user/profile/constants/backendServices';
 
 const user = {
